brand-pr-manager/requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Google Cloud & Firebase
firebase-admin>=6.0.0
google-cloud-firestore>=2.11.0
# google-cloud-aiplatform>=1.32.0
google-adk>=0.1.0
google-cloud-logging>=3.8.0
google-cloud-monitoring>=2.15.0
google-auth>=2.23.0

# Vector Database
pymilvus>=2.2.0
sentence-transformers>=2.2.2

# Async & Networking
httpx>=0.24.0
aiohttp>=3.8.0
aiofiles>=23.0.0
asyncio-pool>=0.6.0

# Caching & Performance
cachetools>=5.0.0
redis>=4.5.0
tenacity>=8.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0
pytz>=2023.3

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
structlog>=23.1.0

# Monitoring & Observability
prometheus-client>=0.17.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-exporter-gcp-monitoring
# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.24.0

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0
